requirements.txt
absl-py==2.3.1
astunparse==1.6.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.3.0
cryptography==44.0.3
cycler==0.12.1
flatbuffers==25.9.23
flwr==1.23.0
fonttools==4.60.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.76.0
grpcio-health-checking==1.62.3
h5py==3.14.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
iterators==0.0.2
joblib==1.5.2
keras==3.10.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.9.4
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
numpy==2.0.2
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
protobuf==4.25.8
pycparser==2.23
pycryptodome==3.23.0
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
rich==13.9.4
scikit-learn==1.6.1
scipy==1.13.1
shellingham==1.5.4
six==1.17.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
tomli==2.3.0
tomli_w==1.2.0
typer==0.12.5
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==2.0.0
zipp==3.23.0
fastapi
uvicorn
python-multipart
pydantic
